Fix Sampling in Settings
Fix Sampling in Settings

Sampling was always enabled and can't be unset.
On Android devices sampling might crash the app.
So, we need to be able to connect to the device with disabled sampling.
Now settings flags are disabled in constructor, so we apply only flags that are set in UI.

---------

Co-authored-by: e.guskov <e.guskov@gaijin.team>